correction attributs des premières classes du modèle

+++ src/main/java/entities/Segment.java
@@ -1,11 +1,11 @@
 package entities;
 
 public class Segment {
-    private int destId;
+    private long destId;
 
     private String streetName;
 
-    private int length;
+    private double length;
 
     public Segment(int destId, String streetName, int length) {
         this.destId = destId;
@@ -22,7 +22,7 @@ public class Segment {
                 '}';
     }
 
-    public int getDestId() {
+    public long getDestId() {
         return destId;
     }
 
@@ -38,7 +38,7 @@ public class Segment {
         this.streetName = streetName;
     }
 
-    public int getLength() {
+    public double getLength() {
         return length;
     }
